+++ packages/cli/src/cms/commands/push.ts
@@ -9,6 +9,7 @@ import { handleReuniteError } from './utils';
 
 import type { OutputFormat } from '@redocly/openapi-core';
 import type { CommandArgs } from '../../wrapper';
+import type { VerifyConfigOptions } from '../../types';
 
 export type PushOptions = {
   apis?: string[];
@@ -29,13 +30,12 @@ export type PushOptions = {
 
   'default-branch': string;
   domain?: string;
-  config?: string;
   'wait-for-deployment'?: boolean;
   'max-execution-time': number;
   'continue-on-deploy-failures'?: boolean;
   verbose?: boolean;
   format?: Extract<OutputFormat, 'stylish'>;
-};
+} & VerifyConfigOptions;
 
 type FileToUpload = { name: string; path: string };