+++ packages/cli/src/commands/join.ts
@@ -23,13 +23,7 @@ import { isObject, isString, keysOf } from '../utils/js-utils';
 import { COMPONENTS, OPENAPI3_METHOD } from './split/types';
 import { crawl, startsWithComponents } from './split';
 
-import type {
-  Oas3Definition,
-  Document,
-  Oas3Tag,
-  Referenced,
-  RuleSeverity,
-} from '@redocly/openapi-core';
+import type { Oas3Definition, Document, Oas3Tag, Referenced } from '@redocly/openapi-core';
 import type { BundleResult } from '@redocly/openapi-core/lib/bundle';
 import type {
   Oas3Parameter,
@@ -38,6 +32,7 @@ import type {
   Oas3_1Definition,
 } from '@redocly/openapi-core/lib/typings/openapi';
 import type { CommandArgs } from '../wrapper';
+import type { VerifyConfigOptions } from '../types';
 
 const Tags = 'tags';
 const xTagGroups = 'x-tagGroups';
@@ -60,9 +55,7 @@ export type JoinOptions = {
   'prefix-components-with-info-prop'?: string;
   'without-x-tag-groups'?: boolean;
   output?: string;
-  config?: string;
-  'lint-config'?: RuleSeverity;
-};
+} & VerifyConfigOptions;
 
 export async function handleJoin({
   argv,