+++ packages/core/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @redocly/openapi-core
 
+## 1.21.1
+
+### Patch Changes
+
+- Updated @redocly/config to v0.10.0.
+
+## 1.21.0
+
 ## 1.20.1
 
 ## 1.20.0